Fix compilation errors on MSVC

+++ src/seabolt/src/bolt/v3.c
@@ -633,7 +633,7 @@ BoltProtocolV3_set_run_cypher_parameter(struct BoltConnection* connection, int32
 {
     struct BoltProtocolV3State* state = BoltProtocolV3_state(connection);
     struct BoltValue* params = BoltMessage_param(state->run_request, 1);
-    BoltDictionary_set_key(params, index, name, name_size);
+    BoltDictionary_set_key(params, index, name, (int32_t)name_size);
     return BoltDictionary_value(params, index);
 }
